Moved the file-findering into its own package.
This :

* Will allow me to add tests.
  * These are not yet present.
* Simplifies the main driver-file
  * Which will be renamed soon.
