+++ implant.go
@@ -11,17 +11,16 @@ package main
 
 import (
 	"bytes"
-	"compress/gzip"
-	"encoding/hex"
 	"flag"
 	"fmt"
 	"io/ioutil"
 	"os"
 	"os/exec"
-	"path/filepath"
 	"regexp"
 	"strings"
 	"text/template"
+
+	"github.com/skx/implant/finder"
 )
 
 // Created so that multiple inputs can be accecpted
@@ -75,44 +74,6 @@ var (
 	version = "master/latest"
 )
 
-// Resource is the structure used to temporarily hold data about all the
-// static-resources we've discovered and will write out to the generated
-// `static.go` file.
-type Resource struct {
-
-	// Filename holds the filename of the resource we've discovered.
-	Filename string
-
-	// Contents holds the string contents of the file.
-	Contents string
-
-	// Length contains the length of the input file.
-	Length int
-}
-
-// CheckInput is designed to test that the input path
-// specified is a directory.
-//
-// If this function returns false then the application terminates.
-func CheckInput(directory string) bool {
-
-	stat, err := os.Stat(directory)
-	if err != nil {
-		fmt.Printf("Failed to stat %s - Does it exist?\n", directory)
-		return false
-	}
-
-	//
-	// Test that the input path is a directory.
-	//
-	if !stat.IsDir() {
-		fmt.Printf("Error %s is not a directory!\n", directory)
-		return false
-	}
-
-	return true
-}
-
 //
 // PipeCommand pipes some data through a binary, and returns the output.
 //
@@ -130,134 +91,9 @@ func PipeCommand(cmd string, input []byte) []byte {
 	return output
 }
 
-// ShouldInclude is invoked by our filesystem-walker, and determines whether
-// any particular directory-entry beneath the input tree should be included
-// in our generated `static.go` file.
-//
-// We skip directories, and we might skip particular entries via a user-supplied
-// regular expression too.
-func ShouldInclude(path string) bool {
-
-	//
-	// Examine the file.
-	//
-	stat, err := os.Stat(path)
-	if err != nil {
-		if ConfigOptions.Verbose {
-			fmt.Printf("Failed to stat file : %s\n", path)
-		}
-		return false
-	}
-
-	//
-	// Is it a regular file?
-	//
-	mode := stat.Mode()
-	if !mode.IsRegular() {
-		//
-		// If it isn't we shouldn't include it.
-		//
-		return false
-	}
-
-	//
-	// If we have a regular expression then test against it.
-	//
-	if ConfigOptions.Exclude != "" {
-
-		//
-		// If the regexp matches then we must EXCLUDE
-		// the file.
-		//
-		match, _ := regexp.MatchString(ConfigOptions.Exclude, path)
-		if match {
-			return false
-		}
-	}
-
-	//
-	// OK the file wasn't excluded, so we'll add it.
-	//
-	return true
-}
-
-// findFiles finds all the files in the given directory, returning an array
-// of Resource-structures to describe each one.
-func findFiles(directory string) ([]Resource, error) {
-
-	// The resources we've found.
-	var entries []Resource
-
-	// The list of files we'll add
-	fileList := []string{}
-
-	// Recursively look for files to add.
-	err := filepath.Walk(directory, func(path string, f os.FileInfo, err error) error {
-		if ConfigOptions.Verbose {
-			fmt.Printf("Found file: %s\n", path)
-		}
-		if ShouldInclude(path) {
-			if ConfigOptions.Verbose {
-				fmt.Printf("\tIncluding file: %s\n", path)
-			}
-			fileList = append(fileList, path)
-		} else {
-			if ConfigOptions.Verbose {
-				fmt.Printf("\tExcluding file: %s\n", path)
-			}
-
-		}
-		return nil
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	// For each file we'll now add it to our list.
-	for _, file := range fileList {
-
-		//
-		// The entry for this file
-		//
-		var tmp Resource
-
-		//
-		// Read the file contents.
-		//
-		data, err := ioutil.ReadFile(file)
-		if err != nil {
-			return nil, err
-		}
-
-		//
-		// gzip the data.
-		//
-		var gzipped bytes.Buffer
-		gw, err := gzip.NewWriterLevel(&gzipped, gzip.BestSpeed)
-		if err != nil {
-			return nil, err
-		}
-		_, err = gw.Write(data)
-		if err != nil {
-			return nil, err
-		}
-		gw.Close()
-
-		//
-		// Add the filename + data, which is now encoded
-		// such that it is printable in our template.
-		//
-		tmp.Filename = file
-		tmp.Contents = hex.EncodeToString(gzipped.Bytes())
-		tmp.Length = len(data)
-		entries = append(entries, tmp)
-	}
-	return entries, nil
-}
-
 // renderTemplate populates our template-file with the list of resources
 // which have been discovered.
-func renderTemplate(entries []Resource) (string, error) {
+func renderTemplate(entries []finder.Resource) (string, error) {
 
 	//
 	// Get our template-text.
@@ -276,7 +112,7 @@ func renderTemplate(entries []Resource) (string, error) {
 	// Populate the template-data
 	//
 	var Templatedata struct {
-		Resources []Resource
+		Resources []finder.Resource
 		Package   string
 	}
 	Templatedata.Package = ConfigOptions.Package
@@ -307,50 +143,48 @@ func Implant() {
 	// We'll build up a list of resources which will be
 	// inserted into our template.
 	//
-	var resources []Resource
+	var resources []finder.Resource
 
 	//
 	// Process each directory which was specified.
 	//
 	for _, directory := range ConfigOptions.Input {
 
-		//
-		// If we're running verbosely show our settings.
-		//
 		if ConfigOptions.Verbose {
 			fmt.Printf("Reading input directory %s\n", directory)
 		}
 
 		//
-		// Check we've been given an input path that a) exists and b) is
-		// a directory.
+		// Find the files beneath the input-directory.
 		//
-		if !CheckInput(directory) {
-			os.Exit(1)
-		}
+		helper := finder.New()
+		entries, err := helper.FindFiles(directory)
 
-		//
-		// Now find the files beneath that path.
-		//
-		files, err := findFiles(directory)
 		if err != nil {
-			fmt.Printf("Error processing files: %s\n", err.Error())
-			return
+			fmt.Printf("Failed to find files beneath %s - %s\n", directory, err.Error())
+			continue
 		}
 
 		//
-		// Append the files to our result-set
+		// If we have a regular expression then test against it.
 		//
-		for _, entry := range files {
-			resources = append(resources, entry)
-		}
+		if ConfigOptions.Exclude != "" {
 
-		//
-		// If a given directory resulted in no input then report that.
-		//
-		if len(files) < 1 {
-			fmt.Printf("Failed to find files beneath %s\n", directory)
-			break
+			//
+			// If the regexp matches then we must EXCLUDE
+			// the file.
+			//
+			for _, entry := range entries {
+				match, _ := regexp.MatchString(ConfigOptions.Exclude, entry.Filename)
+				if !match {
+					resources = append(resources, entry)
+				}
+			}
+		} else {
+
+			for _, entry := range entries {
+				resources = append(resources, entry)
+			}
 		}
 	}
